Moving things around for restarting the project
